Industrial_Automation-main/requirements.txt
pymixbox
pytango
PyQt5